Second part of rune parsing, the nasty part.
As these messages are dem_single, they're not sent to the demo view.
This change removes that filter which means that a lot of other messages
also end up in the parse function. Hopefully taking care of most of this.

Might be glitchy, and does not work for some old demos. A start.